Added some nil checks

+++ app/jobs/process_import.rb
@@ -144,6 +144,8 @@ module Jobs
                 read_timeout: 30
             )
 
+            return nil if html_file.nil?
+
             html = html_file.read
             html_file.close
             html_file.unlink
@@ -169,6 +171,8 @@ module Jobs
 
                 # find topic URL in mapping
                 topic_url = wiki_page_topic_map[title]&.url
+
+                next if topic_url.nil?
 
                 # fix any URLs without a topic title
                 topic_url.sub! '/t//', '/t/'
